:sparkles:  feat: state에 따른 이슈 목록 개수를 구하는 기능 구현

+++ BE/issue-tracker/src/main/java/codesquad/issuetracker/label/LabelRepository.java
@@ -18,7 +18,7 @@ public interface LabelRepository extends CrudRepository<Label, Long>, LabelCusto
     Optional<Label> findById(Long id);
 
     @Query("SELECT COUNT(*) FROM LABEL WHERE IS_DELETED = FALSE")
-    Long countLabels();
+    int countLabels();
 
     @Modifying
     @Query("UPDATE LABEL SET IS_DELETED = TRUE WHERE ID = :labelId")
